TaskManger/requirements.txt
asgiref==3.8.1
coverage==7.6.0
Django==5.0.7
djangorestframework==3.15.2
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
psycopg==3.2.1
psycopg2-binary==2.9.9
pytest==8.3.1
pytest-cov==5.0.0
pytest-django==4.8.0
sqlparse==0.5.1
typing_extensions==4.12.2
